Moved the datastore inside baselib [demos]
