+++ openquake/baselib/datastore.py
